+++ pkg/providers/cloudcertificates/testdata/active_custom_edgerc_section/variables.tf
@@ -1,0 +1,9 @@
+variable "edgerc_path" {
+  type    = string
+  default = "custom_edgerc_path"
+}
+
+variable "config_section" {
+  type    = string
+  default = "custom-section"
+}