+++ pkg/providers/domainownership/testdata/basic_not_validated/import.sh
@@ -1,0 +1,2 @@
+terraform init
+terraform import akamai_property_domainownership_domains.example-com "example.com:DOMAIN,sub.example.com:HOST"